feat: compile large circuit outside tests
