+++ std/compress/lzss_v2/snark_testing.go
@@ -1,6 +1,19 @@
 package lzss_v2
 
-import "github.com/consensys/gnark/frontend"
+import (
+	"fmt"
+	"github.com/consensys/gnark-crypto/ecc"
+	"github.com/consensys/gnark-crypto/ecc/bn254/fr"
+	"github.com/consensys/gnark-crypto/hash"
+	"github.com/consensys/gnark/constraint"
+	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/frontend/cs/scs"
+	"github.com/consensys/gnark/profile"
+	"github.com/consensys/gnark/std/compress"
+	"github.com/consensys/gnark/std/hash/mimc"
+	"os"
+	"time"
+)
 
 type DecompressionTestCircuit struct {
 	C                []frontend.Variable
@@ -25,5 +38,131 @@ func (c *DecompressionTestCircuit) Define(api frontend.API) error {
 			api.AssertIsEqual(c.D[i], dBack[i])
 		}
 	}
+	return nil
+}
+
+func BenchCompressionE2ECompilation(dict []byte, name string) (constraint.ConstraintSystem, error) {
+	d, err := os.ReadFile(name + "/data.bin")
+	if err != nil {
+		return nil, err
+	}
+
+	// compress
+
+	compressor, err := NewCompressor(dict)
+	if err != nil {
+		return nil, err
+	}
+
+	c, err := compressor.Compress(d)
+	if err != nil {
+		return nil, err
+	}
+
+	cStream := ReadIntoStream(c, dict)
+
+	/*cSum, err := check(cStream, cStream.Len())
+	if err != nil {
+		return nil, nerr
+	}
+
+	_, err = check(compress.NewStreamFromBytes(d), len(d))
+	if err != nil {
+		return nil, err
+	}*/
+
+	circuit := compressionCircuit{
+		C:    make([]frontend.Variable, cStream.Len()),
+		D:    make([]frontend.Variable, len(d)),
+		Dict: make([]byte, len(dict)),
+	}
+
+	var start int64
+	resetTimer := func() {
+		end := time.Now().UnixMilli()
+		if start != 0 {
+			fmt.Println(end-start, "ms")
+		}
+		start = end
+	}
+
+	// compilation
+	fmt.Println("compilation")
+	p := profile.Start()
+	resetTimer()
+	cs, err := frontend.Compile(ecc.BN254.ScalarField(), scs.NewBuilder, &circuit)
+	if err != nil {
+		return nil, err
+	}
+	p.Stop()
+	fmt.Println(1+len(d)/1024, "KB:", p.NbConstraints(), "constraints, estimated", (p.NbConstraints()*600000)/len(d), "constraints for 600KB at", float64(p.NbConstraints())/float64(len(d)), "constraints per uncompressed byte")
+	resetTimer()
+	err = compress.GzWrite("../test_cases/"+name+"/e2e_cs.gz", cs)
+	return cs, err
+
+	// setup
+	/*fmt.Println("kzg setup")
+	kzgSrs, err := test.NewKZGSRS(cs)
+	resetTimer()
+	fmt.Println("plonk setup")
+	assert.NoError(t, err)
+	_, _, err = plonk.Setup(cs, kzgSrs)
+	assert.NoError(t, err)
+	resetTimer()*/
+}
+
+type compressionCircuit struct {
+	CChecksum, DChecksum frontend.Variable `gnark:",public"`
+	C                    []frontend.Variable
+	D                    []frontend.Variable
+	Dict                 []byte
+	CLen, DLen           frontend.Variable
+}
+
+func (c *compressionCircuit) Define(api frontend.API) error {
+
+	fmt.Println("packing")
+	cPacked := compress.Pack(api, c.C, int(compress.Gcd(8, longBackRefType.nbBitsAddress, longBackRefType.nbBitsLength, shortBackRefType.nbBitsAddress, shortBackRefType.nbBitsLength)))
+	dPacked := compress.Pack(api, c.D, 8)
+
+	fmt.Println("computing checksum")
+	if err := checkSnark(api, cPacked, c.CLen, c.CChecksum); err != nil {
+		return err
+	}
+	if err := checkSnark(api, dPacked, c.DLen, c.DChecksum); err != nil {
+		return err
+	}
+
+	fmt.Println("decompressing")
+	dComputed := make([]frontend.Variable, len(c.D))
+	if dComputedLen, err := Decompress(api, c.C, c.CLen, dComputed, c.Dict); err != nil {
+		return err
+	} else {
+		api.AssertIsEqual(dComputedLen, c.DLen)
+		for i := range c.D {
+			api.AssertIsEqual(c.D[i], dComputed[i]) // could do this much more efficiently in groth16 using packing :(
+		}
+	}
+
+	return nil
+}
+
+func check(s compress.Stream, padTo int) (checksum fr.Element, err error) {
+
+	s.D = append(s.D, make([]int, padTo-len(s.D))...)
+
+	csb := s.Checksum(hash.MIMC_BN254.New(), fr.Bits)
+	checksum.SetBytes(csb)
+	return
+}
+
+func checkSnark(api frontend.API, e []frontend.Variable, eLen, checksum frontend.Variable) error {
+	hsh, err := mimc.NewMiMC(api)
+	if err != nil {
+		return err
+	}
+	hsh.Write(e...)
+	hsh.Write(eLen)
+	api.AssertIsEqual(hsh.Sum(), checksum)
 	return nil
 }
